# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Setu Health API",
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "paths": {
...
    "/v1/vitals/worklist": {
      "get": {
        "responses": {
          "200": {
            "description": "today's waiting visits at the branch",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "day": {
                      "type": "string"
                    },
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "token": {
                            "type": "string"
                          },
                          "day": {
                            "type": "string"
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "planned",
                              "arrived",
                              "triaged",
                              "in-progress",
                              "finished",
                              "cancelled",
                              "entered-in-error"
                            ]
                          },
                          "patient": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "facilityNo": {
                                "type": "string"
                              },
                              "nameBn": {
                                "type": "string"
                              },
                              "nameEn": {
                                "type": [
                                  "string",
                                  "null"
                                ]
                              },
                              "sex": {
                                "type": "string",
                                "enum": [
                                  "female",
                                  "male",
                                  "other"
                                ]
                              },
                              "birthDate": {
                                "type": [
                                  "string",
                                  "null"
                                ]
                              },
                              "approxAgeYears": {
                                "type": [
                                  "integer",
                                  "null"
                                ]
                              },
                              "approxAgeMonths": {
                                "type": [
                                  "integer",
                                  "null"
                                ]
                              },
                              "approxAgeAt": {
                                "type": [
                                  "string",
                                  "null"
                                ]
                              },
                              "identityConfidence": {
                                "type": "string",
                                "enum": [
                                  "verified",
                                  "unverified",
                                  "possible-duplicate",
                                  "provisional"
                                ]
                              }
                            },
                            "required": [
                              "id",
                              "facilityNo",
                              "nameBn",
                              "nameEn",
                              "sex",
                              "birthDate",
                              "approxAgeYears",
                              "approxAgeMonths",
                              "approxAgeAt",
                              "identityConfidence"
                            ]
                          },
                          "hasVitals": {
                            "type": "boolean"
                          }
                        },
                        "required": [
                          "id",
                          "token",
                          "day",
                          "status",
                          "patient",
                          "hasVitals"
                        ]
                      }
                    }
                  },
                  "required": [
                    "day",
                    "items"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message_bn": {
                      "type": "string"
                    },
                    "message_en": {
                      "type": "string"
                    },
                    "field": {
                      "type": "string"
                    },
                    "reason": {
                      "type": "string"
                    },
                    "canRequest": {
                      "type": "boolean"
                    },
                    "fields": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "code": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "field",
                          "code"
                        ]
                      }
                    },
                    "existing": {
                      "type": "object",
                      "additionalProperties": {}
                    }
                  },
                  "required": [
                    "code",
                    "message_bn",
                    "message_en"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}